fix: extract duplicate code to helper.py
